[feat] Implement drain endpoint
To be used in conjunction with k8s `lifecycle.preStop.httpGet` hook to
configure graceful draining of the pods.  Once the HTTP request is
received at DRAIN_ADDR, the main HTTP server will start reporting
un-healthy (HTTP 500) on `/health` endpoint.  This will signal the
load-balancer that this pod should be drained.